move k8s out of pkg
