+++ cmd/calyptia/create_aggregator_k8s.go
@@ -3,7 +3,6 @@ package main
 import (
 	"context"
 	"fmt"
-	"github.com/calyptia/cli/pkg/k8s"
 
 	"github.com/spf13/cobra"
 	apiv1 "k8s.io/api/core/v1"
@@ -11,6 +10,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 
 	cloud "github.com/calyptia/api/types"
+	"github.com/calyptia/cli/k8s"
 )
 
 const (
